V2.2-REFACTOR - StatusCode Implementation
