+++ src/main/java/co/com/bancodebogota/api/ReqresController.java
@@ -1,9 +1,8 @@
 package co.com.bancodebogota.api;
 
 
-import co.com.bancodebogota.context.account.limits.application.reqres.DataReqres;
+import co.com.bancodebogota.context.account.limits.application.data.Information;
 import co.com.bancodebogota.context.account.limits.domain.user.DataUser;
-import co.com.bancodebogota.context.account.limits.domain.user.DataUsers;
 import io.micronaut.http.annotation.Controller;
 import io.micronaut.http.annotation.Get;
 import io.micronaut.http.annotation.PathVariable;
@@ -12,33 +11,22 @@ import io.micronaut.http.annotation.Post;
 @Controller("/api/${api.version:v1}")
 public final class ReqresController {
 
-    private final DataReqres dataReqres;
+    private final Information information;
 
-    public ReqresController(DataReqres dataReqres) {
+    public ReqresController(Information information) {
 
-        this.dataReqres = dataReqres;
+        this.information = information;
 
-    }
-
-    @Get(value = "users/apimyself")
-    public Object getApiMyself(){
-        return dataReqres.myselfApi();
     }
 
     @Get(value="users/{iduser}")
     public DataUser GetUser(@PathVariable("iduser") int idUser) {
-        return dataReqres.byId(idUser);
+        return information.byId(idUser);
     }
 
-    @Get(value = "users/list/{idPage}")
-    public DataUsers GetUsersList(@PathVariable("idPage")int idPage){
-        return dataReqres.byPage(idPage);
+    @Post(value = "users/create")
+    public Object CreateUser(){
+        return information.create();
     }
-
-   @Post(value = "users")
-    public Object createUser(){
-        return dataReqres.create();
-    }
-
 
 }
