+++ src/main/java/co/com/bancodebogota/client/ReqresClient.java
@@ -8,24 +8,6 @@ import io.micronaut.http.client.annotation.Client;
 
 public interface ReqresClient {
 
-    Object myselfApi();
-
     DataUser byId(@PathVariable("iduser") int idUser);
-
-
-    DataUsers byPage(@PathVariable("idPage")int idPage);
-
     Object create();
-
-
-    //HttpResponse createUser();
-
-    /*
-    HttpResponse updateUser(@PathVariable("idUser") int idUser);
-
-
-    HttpResponse deleteUser(@PathVariable("idUser") int idUser);*/
-
-
-
 }
