+++ src/main/java/co/com/bancodebogota/context/account/limits/application/data/Information.java
@@ -1,24 +1,17 @@
-package co.com.bancodebogota.context.account.limits.application.reqres;
+package co.com.bancodebogota.context.account.limits.application.data;
 
 import co.com.bancodebogota.client.ReqresClient;
 import co.com.bancodebogota.context.account.limits.domain.interfaces.UserRepository;
 import co.com.bancodebogota.context.account.limits.domain.user.DataUser;
-import co.com.bancodebogota.context.account.limits.domain.user.DataUsers;
 import jakarta.inject.Singleton;
 
 @Singleton
-public class DataReqres implements ReqresClient {
+public class Information implements ReqresClient {
 
     private final UserRepository userRepository;
 
-    public DataReqres(UserRepository userRepository) {
+    public Information(UserRepository userRepository) {
         this.userRepository = userRepository;
-    }
-
-
-    @Override
-    public Object myselfApi() {
-        return userRepository.apiMyself();
     }
 
     @Override
@@ -27,12 +20,8 @@ public class DataReqres implements ReqresClient {
     }
 
     @Override
-    public DataUsers byPage(int idPage) {
-        return userRepository.getUsersListData(idPage);
+    public Object create() {
+        return userRepository.createUser();
     }
 
-    @Override
-    public Object create() {
-        return userRepository.createUserReqres();
-    }
 }
