--- src/main/java/co/com/bancodebogota/context/account/limits/domain/errors/Message.java
@@ -1,20 +1,0 @@
-package co.com.bancodebogota.context.account.limits.domain.errors;
-
-public class Message {
-    private String message;
-
-    public Message() {
-    }
-
-    public Message(String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-}
