+++ src/main/java/co/com/bancodebogota/context/account/limits/domain/user/DataUser.java
@@ -1,14 +1,11 @@
 package co.com.bancodebogota.context.account.limits.domain.user;
 
 import io.micronaut.core.annotation.Introspected;
+import io.micronaut.http.HttpResponse;
+import io.micronaut.http.HttpStatus;
 
 @Introspected
 public class DataUser{
     public Data data;
     public Support support;
-    public String message;
-
-    public DataUser(String message) {
-        this.message = message;
-    }
 }
