+++ src/main/resources/application.yml
@@ -13,7 +13,7 @@ micronaut:
         num-threads: 10
 api:
   regres:
-    url: http://localhost:8080/api/bdb
+    url: https://reqres.in/api
